Update types imports
