+++ src/components/atoms/avatar/Avatar.module.css
@@ -1,12 +1,14 @@
-.container { 
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    background-color: var(--avatar-bg-color, #ccc);
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);   
-    transition: background-color 0.3s, box-shadow 0.3s; 
-    font-weight: bold;
+.container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  background-color: var(--avatar-bg-color, #ccc);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  transition:
+    background-color 0.3s,
+    box-shadow 0.3s;
+  font-weight: bold;
 }
