+++ src/components/atoms/avatar/Avatar.stories.tsx
@@ -3,8 +3,7 @@ import Avatar from './Avatar';
 const meta: Meta<typeof Avatar> = {
   title: 'Atoms/Avatar',
   component: Avatar,
-  argTypes: {
-  }
+  argTypes: {}
 };
 
 export default meta;
