--- src/components/atoms/avatar/Avatar.types.ts
@@ -1,3 +1,0 @@
-export interface AvatarProps {
-  name: string;
-}
