--- src/components/atoms/button/Button.types.ts
@@ -1,7 +1,0 @@
-export interface ButtonPropsTypes {
-  children: React.ReactNode;
-  onClick?: () => void;
-  className?: string;
-  disabled?: boolean;
-  variant?: 'primary' | 'secondary';
-}
