+++ src/components/atoms/chip/Chip.tsx
@@ -1,5 +1,10 @@
 import styles from './Chip.module.css';
-import { ChipProps } from './Chip.types';
+
+export interface ChipProps {
+  label: string;
+  className?: string;
+}
+
 const Chip: React.FC<ChipProps> = ({ className, label }) => {
   return (
     <div className={`${styles.chip} ${className}`}>
