--- src/components/atoms/chip/Chip.types.ts
@@ -1,4 +1,0 @@
-export interface ChipProps {
-  label: string;
-  className?: string;
-}
