+++ src/components/atoms/dropdown/Dropdown.tsx
@@ -1,8 +1,29 @@
 import styles from './Dropdown.module.css';
-import { DropdownProps, DropdownType } from './Dropdown.types';
 import { countries, getEmojiFlag } from 'countries-list';
 import type { TCountryCode } from 'countries-list';
 import { v4 as uuid } from 'uuid';
+
+export interface DropdownProps {
+  options?: string[];
+  type?: DropdownType;
+  className?: string;
+  label?: string;
+  onChange: (value: string) => void;
+}
+
+// Add a type-level check to ensure at least one of 'options' or 'type' is provided
+type RequireOptionsOrType<
+  T extends { options?: string[]; type?: DropdownType }
+> =
+  | (T & { options: string[]; type?: never })
+  | (T & { options?: never; type: DropdownType });
+
+export type StrictDropdownProps = RequireOptionsOrType<DropdownProps>;
+
+export enum DropdownType {
+  country = 'country'
+}
+
 const Dropdown: React.FC<DropdownProps> = ({
   options,
   className = '',
