--- src/components/atoms/textfield/Textfield.types.ts
@@ -1,9 +1,0 @@
-export interface TextfieldProps {
-  label?: string;
-  placeholder?: string;
-  value?: string;
-  disabled?: boolean;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  className?: string;
-  type?: 'text' | 'password';
-}
