+++ tsconfig.json
@@ -3,17 +3,28 @@
     "target": "ES2020",
     "module": "ESNext",
     "lib": ["DOM", "ESNext"],
-    "declaration": true,
-    "declarationMap": true,
-    "outDir": "dist/",
-        "rootDir": "src",
     "moduleResolution": "Node",
-    "jsx": "react-jsx", 
-    "strict": true,
+    "jsx": "react-jsx",
+    "declaration": true,
+    "declarationMap": false,
+    "emitDeclarationOnly": false,
+    "outDir": "dist",
+    "rootDir": "src",
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "noEmit": false
   },
   "include": ["src", "global.d.ts"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": [
+    "node_modules",
+    "dist",
+    "**/*.test.ts",
+    "**/*.test.tsx",
+    "**/*.spec.ts",
+    "**/*.spec.tsx",
+    "**/*.stories.ts",
+    "**/*.stories.tsx"
+  ]
 }
